Updated tables; added classes; added search bar for users

+++ ReenbitMessenger.Database/Tables/GroupChatMembers.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[GroupChatMembers]
 (
-	[Id] INT NOT NULL PRIMARY KEY IDENTITY, 
+	[Id] BIGINT NOT NULL PRIMARY KEY IDENTITY, 
     [GroupChatId] CHAR(36) NOT NULL, 
     [UserId] NVARCHAR (450) NOT NULL, 
     [GroupChatRoleId] TINYINT NOT NULL,
